add python version

+++ telebot/modules/sql/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, ARRAY, Integer
+from sqlalchemy import Column, String, ARRAY, BIGINT
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -8,4 +8,4 @@ class CommandExceptionGroups(Base):
     __tablename__ = "CommandExceptionGroups"
 
     command = Column(String, primary_key=True)
-    groups = Column(ARRAY(Integer))
+    groups = Column(ARRAY(BIGINT))
